+++ dashboard/src/components/device-pins/device-pins.tsx
