+++ dashboard/src/index.scss
@@ -16,9 +16,9 @@
 @import './css/tag.scss';
 
 @import './pages/dashboard/dashboard.scss';
-@import './pages/devices/devices.scss';
 @import './pages/device/device.scss';
-@import './pages/device-pin-data/device-pin-data.scss';
+@import './pages/device-view/device-view.scss';
+@import './pages/device-form/device-form.scss';
 @import './pages/oauth-clients/oauth-clients.scss';
 @import './pages/oauth-client/oauth-client.scss';
 @import './pages/triggers/triggers.scss';
