+++ dashboard/src/index.tsx
@@ -3,9 +3,9 @@ import './index.scss';
 import App from './app';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Dashboard from './pages/dashboard/dashboard';
-import DeviceView from './pages/devices/devices';
+import DeviceView from './pages/device-view/device-view';
 import Device from './pages/device/device';
-import DeviceForm from './pages/device-pin-data/device-pin-data';
+import DeviceForm from './pages/device-form/device-form';
 import NotFound from './pages/not-found/not-found';
 import Triggers from './pages/triggers/triggers';
 import Trigger from './pages/trigger/trigger';
