+++ dashboard/src/pages/device-form/device-form.scss
@@ -4,7 +4,7 @@ $pin-option2-h: 80px;
 $pin-option4-h: 160px;
 $dialog-w: 300px;
 
-.device-pin-data {
+.device-form {
     .microcontroller-img {
         max-width: 500px;
     }
