+++ dashboard/src/pages/device-view/device-view.scss
