+++ dashboard/src/pages/device-view/devices.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import DeviceView from './devices';
+import DeviceView from './device-view';
 
 describe('<Devices />', () => {
     test('it should mount', () => {
